% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Artiodactyl}
\alias{Artiodactyl}
\title{Example Dataset: Artiodactyl}
\format{
\subsection{\code{Artiodactyl}}{

A \code{data.frame} of categorical data for 17 species.
\describe{
\item{tip_label}{Common names of 17 artiodactyl species linked to the tree.}
\item{multistate}{Categorical data column.}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Artiodactyl
}
\description{
A dataset used for multi-state analysis.
}
\keyword{datasets}
