% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Artiodactyl_trees}
\alias{Artiodactyl_trees}
\title{Example Tree: Artiodactyl_trees}
\format{
\subsection{\code{Artiodactyl_trees}}{

A \code{multiPhylo} format sample of 500 phylogenetic trees with 17 tips:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Artiodactyl_trees
}
\description{
A sample of 500 trees spanning 17 artiodactyl taxa. Used in some of the BayesTraits example analyses.
}
\keyword{datasets}
