% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArtiodactylMLIn}
\alias{ArtiodactylMLIn}
\title{Example Dataset: ArtiodactylMLIn
An input file used for multi-state analysis.}
\format{
\subsection{\code{ArtiodactylMLIn}}{

An input file used to run multistate analyses.
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
ArtiodactylMLIn
}
\description{
Example Dataset: ArtiodactylMLIn
An input file used for multi-state analysis.
}
\keyword{datasets}
