% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.bt_post.R
\name{autoplot.bt_post}
\alias{autoplot.bt_post}
\title{An autoplot (ggplot2) method for objects of class "bt_post".
Plots a histogram of all, or a specified number of, parameters
from the posterior.}
\usage{
\method{autoplot}{bt_post}(posterior, parameters = NULL, col = NULL,
  pal = "D")
}
\arguments{
\item{posterior}{An object of class "bt_post", returned from
loadPosterior.}

\item{col}{The colour of the bars in the histogram, or 
a vector of colours (for multiple plots). If absent the 
viridis colour scale is used.}

\item{pal}{If using the viridis pallettes (i.e. no user-specified
colours) then this determines the viridis palletter used.
See \link[viridis]{scale_color_viridis}. 
Four options are available: "magma" (or "A"), "inferno" (or "B"), 
"plasma" (or "C"), and "viridis" (or "D", the default option).}

\item{parameter}{The parameter, or vector of parameter names, to plot. 
Must match column names of posterior. If NULL then all parameters 
will be plotted.}
}
\description{
An autoplot (ggplot2) method for objects of class "bt_post".
Plots a histogram of all, or a specified number of, parameters
from the posterior.
}
