% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{BFCrit}
\alias{BFCrit}
\title{Bayes Factor Support.}
\usage{
BFCrit(BF)
}
\arguments{
\item{BF}{a Bayes Factor.}
}
\value{
Verbose support level for Bayes Factor tests.
}
\description{
Given a Bayes Factor, what is the support level from Raftery et al?
}
