% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bird_tree}
\alias{Bird_tree}
\title{Example Tree: Bird_tree}
\format{
\subsection{\code{Bird_tree}}{

A \code{Phylo} format phylogenetic tree of birds with 5669 species:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Bird_tree
}
\description{
A phylogeny of 5669 bird species.
}
\keyword{datasets}
