% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BirdHetCom}
\alias{BirdHetCom}
\title{Example Dataset: BirdHetCom
An input file used to run a heterogeneous model.}
\format{
\subsection{\code{BirdHetCom}}{

An input file used to run a heterogeneous model, where a different evolutionary model is fitted to the Passeriformes.
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
BirdHetCom
}
\description{
Example Dataset: BirdHetCom
An input file used to run a heterogeneous model.
}
\keyword{datasets}
