% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BirdTerritory}
\alias{BirdTerritory}
\title{Example Dataset: BirdTerritory
An input data file used to run a heterogeneous model.}
\format{
\subsection{\code{BirdTerritory}}{

A \code{data.frame} of binary data for birds used to fit a heterogeneous model, usinga lard bird phylogeny.
\describe{
\item{tip_label}{Species names of 5669 artiodactyl species linked to the tree.}
\item{Territory}{Binary data column.}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
BirdTerritory
}
\description{
Example Dataset: BirdTerritory
An input data file used to run a heterogeneous model.
}
\keyword{datasets}
