% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPtools.R
\name{BP_SPblock}
\alias{BP_SPblock}
\title{Generate BayesPhylogenies analysis files.}
\usage{
BP_SPblock(
  nexusfile,
  chainsN = 1,
  model = "GTR",
  iterations = 1100,
  sample = 500,
  burnin = 100,
  seed = "random",
  autorun = T,
  rjpatterns = F,
  gamma,
  blsets,
  rjbranchlengths,
  topologies,
  optionalarguments
)
}
\arguments{
\item{chainsN}{Specify the number of chains to be run. Default value = \code{1}.}

\item{model}{A character string specifying the model of evolution. It must be one of
"GTR" (the default), "GTNR", "HKY85", "F81", "SYM", "JC", "K2P" for sequence data.
For binary data, model can be "M1P", "M2P" and for multi-state data the "KSTATES"
model can be used.}

\item{seed}{A numeric value specifying the random number seed for a model.
If left unspecified, BayesPhylogenies will randomly specify this. If specified,
it is possible to replicate a Markov chain using a given seed.}

\item{autorun}{A Boolean operator. If \code{autorun=TRUE}, then the program
will run immediately without user-input if the output of this function
is executed within BayesPhylogenies.}

\item{rjpatterns}{Boolean operator specifying whether or not to implement the
Pagel and Meade (2004) reversible-jump mixture model for pattern heterogeneity.
Defaults to \code{FALSE}.}

\item{gamma}{Add gamma rate heterogeneity with discrete classes = \code{gamma}.}

\item{blsets}{An optional numeric value that will implement the heterotachy model
as described in (Meade and Pagel, 2008; Pagel and Meade, 2008) for a given
number of branch length sets. Incompatible with \code{rjbranchlengths}.}

\item{rjbranchlengths}{an optional numeric value that implements the reversible-jump
heterotachy model for a given number of branch length sets. Incompatible with
\code{blsets}.}

\item{topologies}{An optional numeric value that allows for multiple topologies
to be estimated. If unspecified, the standard single topology model is run.
It is not recommended that this value be larger than 2.}

\item{optionalarguments}{An optional character vector containing further arguments
to be passed to BayesPhylogenies. Please see program documentation for options
and examples of accepted arguemnts.}

\item{bi}{Numeric value determining the amount of iterations to remove from
the beginning of the chain as burn-in. Defaults to \code{100}.}

\item{it}{Numeric value determining the total number of iterations to run.
Default is \code{1100}.}

\item{sa}{Numeric value determining the "thinning" or "sampling" parameter -
i.e. how many iterations to run between each sample. Default is \code{500}.}
}
\value{
A character string representing the input to BayesTraits
}
\description{
Flexible function for automatically generating the BayesPhylogenies
command block required for running a single-partitioned model.
}
\details{
BayesPhylogenies is in constant development and as such not all eventualities are
accounted for in this package. This function is designed with the informed
user in mind. It is assumed that the user knows which model and parameter
combinations are appropriate for their own data.
}
