% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trees.R
\name{branchdefs}
\alias{branchdefs}
\title{Get all descendants of all nodes in a tree}
\usage{
branchdefs(tree)
}
\arguments{
\item{tree}{a phylogenetic tree as an object of class \code{phylo}.}
}
\value{
A \code{list} of vectors defining the taxa descending from every node of a tree.
}
\description{
Retrieves the descendants of every node (including terminals) in a tree.
}
\keyword{internal}
