% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_code.R
\name{bt_addmodel}
\alias{bt_addmodel}
\title{Create model: Code to determine restrictions for a DAG}
\usage{
bt_addmodel(dag)
}
\arguments{
\item{dag}{a description of the model using DAGs. These can be created at dagitty.net and copied into this function.}
}
\value{
a string of Restrictions that can be placed into a BayesTraits script.
}
\description{
For a given DAG, this function will return the necessary restrictions for that model to be estimated
Currently, this function returns all parameters that must be restricted to zero.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
