% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_code.R
\name{bt_addtag}
\alias{bt_addtag}
\title{AddTag: Code to determine a tag for BayesTraits}
\usage{
bt_addtag(tree, taxa, label = "ANC")
}
\arguments{
\item{tree}{an object of class phylo}

\item{taxa}{a vector of character strings indicating relevant taxa}

\item{label}{the name of the label (ANC by default)}
}
\value{
a string starting with 'AddTag' followed by the label and taxa names
}
\description{
Given a tree and a set of taxa, this function will return all taxa between the two given and return
an AddTag line that can be added to a BayesTraits script to use with other analysis (e.g. RecNode)
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
