% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate.R
\name{bt_read.log}
\alias{bt_read.log}
\title{Read BayesTraits Log files}
\usage{
bt_read.log(filename)
}
\arguments{
\item{filename}{the path to the BayesTraits log file}
}
\value{
A data.frame of the logs found in the BayesTraits log file & a list of settings taken from the header of the file
}
\description{
Given a the path of a BayesTraits log file, this function will find the start of the logged output and import the data as a data.frame into R.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
