% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate.R
\name{bt_read.schedule}
\alias{bt_read.schedule}
\title{Read BayesTraits Schedule files}
\usage{
bt_read.schedule(filename)
}
\arguments{
\item{filename}{the path to the BayesTraits Schedule file}
}
\value{
A data.frame of the schedule found in the BayesTraits schedule file  & a list of settings taken from the header of the file
}
\description{
Given a the path of a BayesTraits Schedule file, this function will find the start of the logged output and import the data as a data.frame into R.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
