% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate.R
\name{bt_read.stones}
\alias{bt_read.stones}
\title{Read BayesTraits Stones files}
\usage{
bt_read.stones(filename)
}
\arguments{
\item{filename}{the path to the BayesTraits Schedule file}
}
\value{
A data.frame of the schedule found in the BayesTraits schedule file  & a list of settings taken from the header of the file
}
\description{
Given a the path of a BayesTraits Stones file, this function will find the start of the stones output and import the data. It will also find the marginal likelihood and return both items in a list.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
