% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{bt_write}
\alias{bt_write}
\title{Write / Create BayesTraits tree and data files}
\usage{
bt_write(tree, data, variables, dir = "./", na.omit = FALSE, filename)
}
\arguments{
\item{tree}{an object of class phylo or list of phylo objects}

\item{data}{a data.frame}

\item{variables}{a character string or character vector of the names of columns to subset to, alternatively you can use the numeric values to indicate the column}

\item{dir}{the directory to save the output files to. Defaults to current directory}

\item{na.omit}{a TRUE/FALSE argument to specify whether NA values should be removed from your final data file}

\item{optional:}{used to name the output file. Will use variable names by default.}
}
\value{
two files will be saved to the specified directory *.bttrees holding the pruned tree files and *.btdata holding the subset and formatted data.
}
\description{
Given a tree & data.frame or vector, this function will create input files for BayesTraits with the suffix \*.bttrees and \*btdata.
This function will also format the data file for BayesTrait \(e.g. remove or replace NA values with "-"\)
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
