% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBTjob.R
\name{BTBetapriors}
\alias{BTBetapriors}
\title{Generate priors for Beta estimation in BayesTraits.}
\usage{
BTBetapriors(data, prior, pars)
}
\arguments{
\item{data}{The data table that is wanted for further analysis in BayesTraits.
This must contain ONLY the columns required for subsequent analysis,
and must have names matching the tree as the first column.}

\item{prior}{Name of the prior to apply. Can only be one of "normal", "unif",
"gamma", "exp". See BayesTraits manual for details.}

\item{pars}{Numeric values defining the limits/parameters of the prior.
See BayesTraits manual for details.}
}
\value{
A character string used in BayesTraits input files to define priors
for a set of Betas. Designed for use with \code{\link[=createBTjob]{createBTjob()}}.
}
\description{
Generate priors formatted for use as an 'optional argument' in \code{\link[=createBTjob]{createBTjob()}}.
This is specifically designed to generate priors for regression betas in order to
avoid typing out variable names manually.
}
