% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trees.R
\name{compareBLS}
\alias{compareBLS}
\title{Compare branch lengths between two trees of identical topology}
\usage{
compareBLS(tree1, tree2)
}
\arguments{
\item{tree1}{a phylogenetic tree as an object of class \code{phylo}.}

\item{tree2}{a phylogenetic tree as an object of class \code{phylo}.}
}
\value{
The set of named tips descended from node in a vector.
}
\description{
A function that extracts branch lengths between two trees of identical topology but where (for whatever reason) the edge matrices of the two trees do not match. This problem is encountered often when calculating stretched trees from the variable rates post-processor and comparing them to  means calculated from the raw trees output by the variable rates model.
}
