% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBTjob.R
\name{createBTjob}
\alias{createBTjob}
\title{Generate BayesTraits analysis files.}
\usage{
createBTjob(
  fm,
  dataset,
  tree,
  jobname,
  bi = 1e+05,
  it = 1100000,
  sa = 1000,
  model,
  MCMC = T,
  reps = 1,
  optarg,
  out,
  DistData,
  link
)
}
\arguments{
\item{fm}{An object of class "formula" that is a symbolic description of the
desired model to be fitted in BayesTraits. See details.}

\item{dataset}{The data table that is wanted for further analysis in BayesTraits.
This must contain ONLY the columns required for subsequent analysis,
and must have names matching the tree as the first column.}

\item{tree}{A tree of class ("Phylo") containing tip-names as found in the first
column of 'dataset'.}

\item{jobname}{A unique character string used for identifying the model.
All created files will have this tag.}

\item{bi}{Numeric value determining the amount of iterations to remove from
the beginning of the chain as burn-in. Defaults to 100000}

\item{it}{Numeric value determining the total number of iterations to run.
Default is 1100000.}

\item{sa}{Numeric value determining the "thinning" or "sampling" parameter -
i.e. how many iterations to run between each sample. Default is 1000.}

\item{model}{A required argument that specifies which model is to be run in
BayesTraits. See BayesTraits manual for list of available models.}

\item{MCMC}{Boolean operator that determines whether the analysis to be run
will be MCMC (T) or ML (F). Defaults to MCMC=TRUE.}

\item{reps}{Numeric value defining the number of replicates of this model to be run.
Defaults to 1.}

\item{optarg}{an optional character vector with arguments to
include in the input file for running BayesTraits. This includes commands
like "varrates" or "Lambda" in addition to prior specifications and
stepping stone sampling. See BayesTraits manual for list of available commands.}

\item{out}{Path to desired folder where all generated job files will be saved.
If this parameter is left undefined, files will be generated in the current
working directory.}

\item{DistData}{Boolean operator that specifies whether there are sampled
values to be accounted for. This allows the user to input a sample of data
for a taxon/range of taxa instead of a single point. This requires the input
data file to contain multiple entries for at least one of the sampled taxa.
See BayesTraits manual for details.}

\item{link}{If DistData = TRUE, then this specifies whether the samples are
linked: where multiple traits within a single taxon are taken from the same source
(e.g. brain and body size taken from a single individual); or unlinked: where
data values for a given taxon are not taken from a single source or individual.
Please see \code{\link[=linkvalues]{linkvalues()}} function and BayesTraits Manual for more details).}
}
\value{
Generates files in the current working directory that can be used
to run BayesTraits analyses.
}
\description{
Flexible function for automatically generating all of the files
required for running phylogenetic statistical comparative analyses in BayesTraits.
}
\details{
BayesTraits is in constant development and as such not all eventualities are
accounted for in this package. This function is designed with the informed
user in mind. It is assumed that the user knows which model and parameter
combinations are appropriate for their own data.
}
\section{Model specification}{

Using this function, models can be specified symbolically. The response
variable must always be a numeric vector. Models can be specified as described
in the documentation for \code{\link[=lm]{lm()}}.
}

