% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{extractML}
\alias{extractML}
\title{Extract marginal likelihood from a stones file}
\usage{
extractML(input)
}
\arguments{
\item{input}{A character vector defining the stones file exactly as output by BayesTraits.}
}
\value{
A numeric value defining the log marginal likelihood of the output model.
}
\description{
Function to extract the marginal likelihood from the raw stepping-stone sampling output.
}
