% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStones.R
\name{getStones}
\alias{getStones}
\title{getStones}
\usage{
getStones(logs, order = TRUE)
}
\arguments{
\item{logs}{A vector of one or more log files from stepping stones analysis.}

\item{order}{If TRUE then the resultant data.frame is ordered according to
marginal likelihood.}
}
\value{
A data.frame with a row per stones file, and a column with the
marginal likelihoods of each model in.
}
\description{
Get the marginal likelihoods from from one or more stepping stones log
files.
}
