% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiags.R
\name{ggHist}
\alias{ggHist}
\title{ggHist}
\usage{
ggHist(posterior, param, title)
}
\arguments{
\item{posterior}{The posterior of a bayestraits analysis (from loadPosterior)}

\item{param}{The parameter to plot a histogram of.}
}
\value{
A histogram with nice ggplot aesthetics, and good bin widths.
}
\description{
A function to plot a nice histogram - one that uses ggplot2
but has sensible break behaviour in R. Produces a single
histogram but is mostly called internally for things like
plotPosts.
}
