% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HarmonicMeanLh}
\alias{HarmonicMeanLh}
\title{Example Dataset: HarmonicMeanLh
An example of 10,000 log likelihoods.}
\format{
\subsection{\code{HarmonicMeanLh}}{

A file containing a list of log-likelihoods that can be used with the web-calculator to estimate the log harmonic mean. However, please note that this functionality has been removed from current versions of BayesTraits in favour of stepping stone sampling.
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
HarmonicMeanLh
}
\description{
Example Dataset: HarmonicMeanLh
An example of 10,000 log likelihoods.
}
\keyword{datasets}
