% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{histRotate}
\alias{histRotate}
\title{Plot rotated histograms.}
\usage{
histRotate(A, direction = "up")
}
\arguments{
\item{A}{An R object obtained from the hist(plot = F) function.}

\item{direction}{The direction for the histogram. Can be any of "up", "down", "left", or "right".}
}
\value{
Plots a rotated histogram.
}
\description{
Allows for plotting of histograms in any orientation: up, down, left, or right.
}
