% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBTjob.R
\name{linkvalues}
\alias{linkvalues}
\title{Generate DistData file}
\usage{
linkvalues(unlinked, link = "Linked")
}
\arguments{
\item{unlinked}{The data table that is wanted for further analysis in BayesTraits.
This MUST contain taxa names or row IDs in the first column.}

\item{link}{Currently, the value must be "Linked" or "Unlinked".
This will be applied across all taxa. Future updates may allow
variable values across taxa.}
}
\value{
An R data frame that can be exported to a text file and used as input
for the DistData command in BayesTraits.This function is used internally by
\code{\link[=createBTjob]{createBTjob()}}.
}
\description{
Takes a data file and generates a DistData file for
analysis in BayesTraits with multiple datapoints for some tips.
}
