% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPosterior.R
\name{loadPosterior}
\alias{loadPosterior}
\title{loadPosterior}
\usage{
loadPosterior(logfile, thinning = 1, burnin = 0)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits analysis.}

\item{thinning}{Thinning parameter for the posterior - defaults to 1
(all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the
chain as burnin. Use if the chain has not reached convergence before sampling
began. Useful if the burnin parameter for the analysis itself was not long
enough.}
}
\value{
A tibble (see \link[tibble]{tibble}) with the class "bt_post" containing
the samples from the BayesTraits MCMC chain. Headers vary on model type.
}
\description{
Returns the full mcmc object from a BayesTraits log file. This
is used inside plot functions and so on, but might be useful for
other MCMC manipulations and so on.
Extracts the MCMC samples from a BayesTraits logfile (i.e. discards the
header information and coerces samples into a matrix.)
}
