% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make colour transparent.}
\usage{
makeTransparent(colour, alpha = 0.5)
}
\arguments{
\item{colour}{Either a named R colour or a hexidecimal code of format "#000000".}

\item{alpha}{The level of desired transparency.}
}
\value{
Hexidecimal code for converted transparent colour.
}
\description{
Convert named colour (or hexadecimal code) to transparent version of the same colour.
}
\examples{
makeTransparent("red", 0.5)
makeTransparent("#000000", 0.1)
}
\author{
Originally taken from \href{https://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color}{stackoverflow}
}
\keyword{internal}
