% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mammal_trees}
\alias{Mammal_trees}
\title{Example Tree: Mammal_trees}
\format{
\subsection{\code{Mammal_trees}}{

A \code{multiPhylo} format sample of 50 phylogenetic trees with 40 tips:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Mammal_trees
}
\description{
A sample of 50 trees spanning 40 mammal taxa. Used in some of the BayesTraits example analyses.
}
\keyword{datasets}
