% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MammalBody}
\alias{MammalBody}
\title{Example Dataset: MammalBody
An input file used for continuous models.}
\format{
\subsection{\code{MammalBody}}{

A \code{data.frame} of body size data for mammals.
\describe{
\item{tip_label}{Species name.}
\item{Body}{logged body size data.}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
MammalBody
}
\description{
Example Dataset: MammalBody
An input file used for continuous models.
}
\keyword{datasets}
