% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MammalBrainBody}
\alias{MammalBrainBody}
\title{Example Dataset: MammalBrainBody
An input file used for corr or reg models.}
\format{
\subsection{\code{MammalBrainBody}}{

A \code{data.frame} of brain and body size data for mammals.
\describe{
\item{tip_label}{Species name.}
\item{Brain}{logged brain size data.}
\item{Body}{logged body size data.}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
MammalBrainBody
}
\description{
Example Dataset: MammalBrainBody
An input file used for corr or reg models.
}
\keyword{datasets}
