% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MammalBrainBodyGt}
\alias{MammalBrainBodyGt}
\title{Example Dataset: MammalBrainBodyGt
An input file used for corr or reg models.}
\format{
\subsection{\code{MammalBrainBodyGt}}{

A \code{data.frame} of body, brain, and gestation time for mammals.
\describe{
\item{tip_label}{Species name.}
\item{Brain}{logged brain size data.}
\item{Body}{logged body size data.}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
MammalBrainBodyGt
}
\description{
Example Dataset: MammalBrainBodyGt
An input file used for corr or reg models.
}
\keyword{datasets}
