% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MammalBrainBodySampleData}
\alias{MammalBrainBodySampleData}
\title{Example Dataset: MammalBrainBodySampleData
An example DistData file}
\format{
\subsection{\code{MammalBrainBodySampleData}}{

An example distData object
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
MammalBrainBodySampleData
}
\description{
Example Dataset: MammalBrainBodySampleData
An example DistData file
}
\keyword{datasets}
