% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MammalModelB}
\alias{MammalModelB}
\title{Example Dataset: MammalModelB
An example trend data file.}
\format{
\subsection{\code{MammalModelB}}{

A \code{data.frame} of fictional data with a trend.
\describe{
\item{tip_label}{Species name.}
\item{trend}{simulated trend data}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
MammalModelB
}
\description{
Example Dataset: MammalModelB
An example trend data file.
}
\keyword{datasets}
