% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marsupial_tree}
\alias{marsupial_tree}
\title{Example Tree: Marsupials}
\format{
\subsection{\code{marsupial_tree}}{

A "phylo" format Phylogenetic tree with 173 tips and 160 internal nodes:
\describe{
\item{Tip Labels}{Binomial as found in the Time Tree of Life, downloaded from timetree.org Jan 2020.}
...
}
}
}
\source{
Downloaded from timetree.org Jan 2020
}
\usage{
marsupial_tree
}
\description{
A subset of the timetree.org mammalian phylogeny representing 173 marsupials
}
\keyword{datasets}
