% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marsupials}
\alias{marsupials}
\title{Example Dataset: Marsupials}
\format{
\subsection{\code{marsupials}}{

A data frame with 173 rows and 6 columns:
\describe{
\item{Species}{Binomial as found in the Time Tree of Life, downloaded from timetree.org Jan 2020.}
\item{Order, Family}{Taxonomic information}
\item{Br,Bd}{Brain and body size data in log10 grams}
\item{Ref}{Source for the data.}
...
}
}
}
\source{
Venditti, Baker & Barton (2023) forthcoming publication
}
\usage{
marsupials
}
\description{
A subset of the mammalian brain size dataset presented in Venditti et al 2023
}
\keyword{datasets}
