% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrixoverlap}
\alias{matrixoverlap}
\title{Pairwise comparison among columns.}
\usage{
matrixoverlap(mat, pmcmc = F)
}
\arguments{
\item{mat}{A matrix of dimension n>1,n>2 with the values of interest. This MUST have column names.}

\item{pmcmc}{If true, instead of returning the proportion of y that is greater than x it will return the proportion of y that is DIFFERENT to x i.e. a pmcmc value (proportion of differences overlapping zero).}
}
\value{
A matrix identifying the proportion of the values that overlap between each pair (pmcmc).
}
\description{
For every pair of values in a matrix, calculates the overlap.
Overlap is the proportional difference between each 2 columns of data. This can be used to identify pairwise difference matrices amongst group-level parameter estimates.
Note that these values have to be paired otherwise this doesn't make much sense!
}
