% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPlots.R
\name{mcmcPlots}
\alias{mcmcPlots}
\title{mcmcPlots}
\usage{
mcmcPlots(logfile, parameters = NULL, ...)
}
\arguments{
\item{logfile}{Either an object of class "bt_post", or the name of the
logfile of the analysis of interest.}

\item{parameters}{A vector of names of parameters to show diagnostic
plots for. If NULL (default) then all parameters are used (depending on
the analysis some of these may be nonsensical).}

\item{...}{Additional arguments passed to loadPosterior}
}
\description{
Plots 4 MCMC diagnostics for a parameter
}
\examples{
paramDiagnostic("cool-data.log", "Lh")
params <- getParams("cool-data.log")
paramDiagnostic("cool-data.log", params, type = "trace", cols = 3)
}
\keyword{autocorrelation}
\keyword{density}
\keyword{diagnostic}
\keyword{mcmc}
\keyword{mean}
\keyword{running}
\keyword{trace}
