% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NortheastBantu}
\alias{NortheastBantu}
\title{Example Dataset: NortheastBantu
An example geo data file.}
\format{
\subsection{\code{NortheastBantu}}{

A \code{data.frame} of Bantu co-ordinates.
\describe{
\item{tip_label}{Species name.}
\item{long}{longitude}
\item{lat}{latitude}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
NortheastBantu
}
\description{
Example Dataset: NortheastBantu
An example geo data file.
}
\keyword{datasets}
