% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NortheastBantu_tree}
\alias{NortheastBantu_tree}
\title{Example Tree: NortheastBantu_tree}
\format{
\subsection{\code{NortheastBantu_tree}}{

A \code{Phylo} format phylogenetic tree of 85 Northeastern bantu languages:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
NortheastBantu_tree
}
\description{
A phylogeny of 85 Northeastern Bantu languages.
}
\keyword{datasets}
