% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlap}
\alias{overlap}
\title{Pairwise comparison between two sets of values.}
\usage{
overlap(x, y, pmcmc = F)
}
\arguments{
\item{x}{A numeric vector with the reference values.}

\item{y}{A numeric vector with the comparison values.}

\item{pmcmc}{If true, instead of returning the proportion of y that is greater than x it will return the proportion of y that is DIFFERENT to x i.e. a pmcmc value (proportion of differences overlapping zero).}
}
\value{
The proportion of the new values that are greater than (or different to, if pmcmc = t) the reference values.
}
\description{
Calculates the difference between 2 columns of data and the proportion of the new values that are greater than the original reference set. Note that these values have to be pairwise otherwise this doesn't make much sense! This can be useful for comparing differences amongst parameter values in MCMC regressions.
}
\author{
Jo Baker
}
