% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trees.R
\name{PL}
\alias{PL}
\title{Calculate path length across a tree.}
\usage{
PL(tree, nodes = F)
}
\arguments{
\item{tree}{A tree of class \code{phylo} used as the reference tree.}

\item{nodes}{Boolean operator (default = \code{FALSE}) that if \code{TRUE} will retain path length estimates for internal nodes as well as terminal taxa.}
}
\value{
A named numeric vector of root-to-tip distances.
}
\description{
Extract the root to tip distances for every terminal taxa (and optionally, internal node) in the tree.
}
