% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.bt_post.R
\name{plot.bt_post}
\alias{plot.bt_post}
\title{A plot method for the class "bt_post" that invokes the 
autoplot method.}
\usage{
\method{plot}{bt_post}(x, ...)
}
\description{
A plot method for the class "bt_post" that invokes the 
autoplot method.
}
