% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{plotBTlog}
\alias{plotBTlog}
\title{Plot parameter values against index value.}
\usage{
plotBTlog(files, cols = "all", colours, burnin = 0)
}
\arguments{
\item{files}{path or filename of the file a summary is required for.
This can be a list of filenames - if so, all identical values in each
table will be plotted on a single chart. This is useful for comparing estimates
of the same parameter across multiple replicate,s for example.
Files must be in the format as output by BayesTraits.}

\item{cols}{A string or list of strings defining the column names or indices
for which summary information is required. If unspecified, takes the
value "all" which will return traces for the following columns
(if present): Likelihood, Intercept, Slope(s)/Beta parameters,
Variance, R-squared, Number of local scalars (branch or node), and Lambda.}

\item{colours}{if provided, then should be a vector of colours for plotting - otherwise will be drawn randomly from rainbow.}

\item{burnin}{A value specifying the number of rows to remove from the beginning of each log file. Default value is 0 and so the entire chain will be plotted.}
}
\description{
Useful for visualizing parameter traces from MCMC analyses in
BayesTraits but can in principle be applied to any data.
}
