% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Primates}
\alias{Primates}
\title{Example Dataset: Primates
An example discrete data file.}
\format{
\subsection{\code{Primates}}{

A \code{data.frame} of binary data for discrete models.
\describe{
\item{tip_label}{Species name.}
\item{trait1}{first binary trait}
\item{trait2}{second binary trait}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Primates
}
\description{
Example Dataset: Primates
An example discrete data file.
}
\keyword{datasets}
