% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Primates_trees}
\alias{Primates_trees}
\title{Example Tree: Primates_trees}
\format{
\subsection{\code{Primates_trees}}{

A \code{multiPhylo} format phylogenetic tree of 60 primate species:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Primates_trees
}
\description{
A sample of 500 phylogenetic trees sampling 60 primate species.
}
\keyword{datasets}
