% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{processVR}
\alias{processVR}
\title{Process BayesTraits VR log file}
\usage{
processVR(VRout)
}
\arguments{
\item{VRout}{An R object defining the output of the read VR function.}
}
\description{
Process the output files from a variable rates analysis in BayesTraits
as read in by readVR.
}
