% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{readVR}
\alias{readVR}
\title{Read in BayesTraits VR Log Files (raw)}
\usage{
readVR(vrfile)
}
\arguments{
\item{vrfile}{Takes a single string defining the direct path to a single log file.}
}
\value{
A list with two elements:
\itemize{
\item brInfo - A \code{data.frame} giving information about each branch/node
in the analysis. Each branch is identified by the numeric value in "branch". For
samples of trees, this value is equivalent to the "partition ID" used internally
by BayesTraits. "NoDes" and "Des" record the number and names of all descendant
taxa from each branch. The "freq" and "probability" columns tell us the frequency
and probability that this branch occur in the sample are only relevant for
analyses over samples of trees. The "length" records the branch length and
is left blank for tree samples.
\item VRlog - A data.frame tabulating the VR output from BayesTraits and
containing the following columns: "It", "Lh", "Lh.Prior", "NoPram", "Alpha",
"Sigma.2", "AlphaScalePrior", "CreateIt", "NodeBranch", "NodeID", "Scalar". If
the original analysis was done on a sample of trees, "NodeID" is replaced by
"Partition ID" and the tree number used at each iteration is recorded.
}
}
\description{
Reads the output file from a variable rates analysis in BayesTraits
for further processing in an R workspace. Will only work with VR log files
(.VarRates.txt) as directly output from BayesTraits models 1-4, 7 and 9.
Will not (currently) work with output from any other program or model.
}
