% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{scaleTree}
\alias{scaleTree}
\title{Create a stretched tree from variable rates output}
\usage{
scaleTree(VRsummary, frequency = 0, magnitude = 1, type = "median")
}
\arguments{
\item{VRsummary}{The name of the object created by \link{summarizeVR}}

\item{frequency}{A percentage ranging between 0-100 specifying which branches to scale. Branches will only be scaled if they have been scaled in a greater percentage of the posterior than is specified by this parameter.This parameter is ignored if type = "sample".}

\item{magnitude}{A value specifying wich branches to scale. Branches will only be scaled if they are above the value specified by magnitude (or below 1/magnitude for rate decreases). This parameter is ignored if type = "sample".}

\item{type}{A character string defining what type of scaled tree to return. Can be one of "median", "mean", or "sample". If sample, the full posterior of scaled trees will be returned.}
}
\value{
A tree where branch lengths are stretched by the mean or median scalar - or a list of trees stretched by the rate scalars.
}
\description{
Creates a stretched tree (mean, median, or full sample) from the output of the variable rates model.
}
