% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoTools.R
\name{sf_to_csv}
\alias{sf_to_csv}
\title{Transform a spatial feature object to csv.}
\usage{
sf_to_csv(
  age = 150,
  grid = NULL,
  extent = c(-180, 180, -90, 90),
  resolution = 2,
  model = "PALEOMAP",
  outputfile = "./masked.csv"
)
}
\arguments{
\item{age}{Target age in million years. Default 150. Based on the \code{\link[rgplates]{reconstruct}} function.}

\item{grid}{The spatial grid to create a mask of land/ocean.}

\item{extent}{A numeric vector defining the mask bounding box. Default is the global extent c(-180, 180, -90, 90). More information about this can be found in the help pages for the \code{\link[terra]{ext}} function from the \code{terra} package.}

\item{resolution}{A numeric value defining the grid cell size in degrees for the WGS84 crs. Default is 2 degrees. There are no explicit bounds for this but values of <0.01 can result in slow compute times.}

\item{model}{The map reconstruction model. Default is "PALEOMAP". Based on the \code{\link[rgplates]{reconstruct}} function - see full list of options there.}

\item{outputfile}{A character string defining the filename (and full filepath) where the csv file should be saved. By default, the file will be saved simply as "masked.csv" in the present working directory. If this argument is left blank, no csv will be generated.}
}
\value{
A \code{data frame} with longitude, latitude, and mask value for land/ocean. This table is also written to the filepath specified by \code{outputfile}.
}
\description{
Get palaeomaps in .csv format to create binary map files in BayesTraits.
}
\examples{
df <- sf_to_csv(age = 200)
df <- sf_to_csv(age = 200, extent = c(-80, 30, -90, 90), resolution = 1)

}
\author{
Jorge Avaria-Llautureo (j.l.avaria@reading.ac.uk).
}
