% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateLogs.R
\name{summarizeBT}
\alias{summarizeBT}
\title{Function to summarize BayesTraits log files}
\usage{
summarizeBT(
  file,
  cols = "all",
  tradeoffs = T,
  table = T,
  name = "Summary",
  input
)
}
\arguments{
\item{file}{data.frame or file path of the BayesTraits output.
This must be one of either (see also the 'table' argument):
- a file path for a tab-delimited text file specifying the output columns from a BayesTraits run.
- a name of an  R data-frame.}

\item{cols}{A string or list of strings defining the columns for which summary information is required.
If unspecified, takes the value "all" which summarizes a small pre-defined list of columns including:
Likelihood, Alpha, Beta estimates, Variance, R-squared, Local transforms, and Lambda.}

\item{tradeoffs}{Boolean operator that if true, plots trade-offs between all parameters specified in cols.}

\item{table}{Boolean operator that if true, accepts an R data.frame object as input. If False, the file argument is interpreted as a character-string file path to the trimmed log file output from BayesTraits.}

\item{name}{The name which the output will be saved under.}

\item{input}{If true, and table = FALSE, then will modify the column names of the output table to reflect the input data file (e.g. Beta.1 will become "BodyMass").}
}
\value{
R data.frame object with the summary information:
\item{Parameter}{Name of the parameter of interest}
\item{Minimum}{Minimum parameter value}
\item{Mean}{Mean parameter value}
\item{Median}{Median parameter value}
\item{Mode}{Mode parameter value}
\item{Maximum}{Maximum parameter value}
\item{PropLess0}{The proportion of the posterior distribution of the parameter that is less than zero.}
\item{pmcmc}{The proportion of the posterior distribution of the parameter that crosses zero i.e. in either direction.}
}
\description{
This function will produce average parameter values (mean, median, mode) as well as other summary statistics.
It will return averages, effective sample sizes, pMCMC values, and ranges.
It ideally works with the output from trimmedlogs().
Can work with all sorts of outputs, but will need to manually specify column names for non Bayes Traits tables.
}
