% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{summarizeBTlog}
\alias{summarizeBTlog}
\title{Function to summarize BayesTraits log files}
\usage{
summarizeBTlog(file, cols = "all", tradeoffs = F, input = T, burnin = 0)
}
\arguments{
\item{file}{data.frame or file path of the BayesTraits output.}

\item{cols}{A string or list of strings defining the columns for which summary information is required.
If unspecified, takes the value "all" which summarizes a small pre-defined list of columns including:
Likelihood, Alpha, Beta estimates, Variance, R-squared, Local transforms, and Lambda.}

\item{tradeoffs}{Boolean operator that if true, plots trade-offs between all parameters specified in cols. Defaults to FALSE, as can be slow for complex models.}

\item{input}{Optional parameter. Can take two values. If simply specified as TRUE, this will search for a file with the name as follows: gsub(".Log.txt", "", file). Otherwise, can take a specified input file name. In either case, when specified, the function will link the column names of the output table to reflect the parameters in the input data file (e.g. Beta.1 is associated with "Body_Mass"). Note that if the order of columns of the output or input files have been modified in any way this will not produce desirable output.}

\item{burnin}{A value specifying the number of rows to remove from the beginning of each log file. Default value is 0 and so the entire chain will be used}
}
\value{
R data.frame object with the summary information:
\item{Parameter}{Name of the parameter of interest}
\item{Minimum}{Minimum parameter value}
\item{Mean}{Mean parameter value}
\item{Median}{Median parameter value}
\item{Mode}{Mode parameter value}
\item{Maximum}{Maximum parameter value}
\item{PropLess0}{The proportion of the posterior distribution of the parameter that is less than zero.}
\item{pmcmc}{The proportion of the posterior distribution of the parameter that crosses zero i.e. in either direction.}
}
\description{
This function will produce average parameter values (mean, median, mode) as well as other summary statistics.
It will return averages, effective sample sizes, pMCMC values, and ranges.
}
