% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{summarizeVR}
\alias{summarizeVR}
\title{Summarize BayesTraits VR Log Files (from raw output)}
\usage{
summarizeVR(vrfile)
}
\arguments{
\item{vrfile}{Takes a single string defining the direct path to a single log file.}
}
\value{
A list with two elements:
\itemize{
\item summary - A \code{data.frame} giving information about each branch/node
in the analysis. Each branch is identified by the numeric value in "branch". For
samples of trees, this value is equivalent to the "partition ID" used internally
by BayesTraits. "NoDes" and "Des" record the number and names of all descendant
taxa from each branch. The "freq" and "probability" columns tell us the frequency
and probability that this branch occur in the sample are only relevant for
analyses over samples of trees. The "length" records the branch length and
is left blank for tree samples. The "DesBranches" tells us all branches that descend from
the branch of interest. The median and mean scalar are the median and mean rate
scalars acting on that branch across the entire posterior distribution. n_scaled
tells us how many iterations that branch had a scalar acting on it, and n_origin tells
us how many iterations that branch had a scalar ORIGINATING on it (either as the base of
a node scalar or as its own branch scalar).
\item posterior - A \code{data.frame} providing the posterior distribution of rate scalars acting
on every branch (column) at each iteration (row).
}
}
\description{
Reads the output file from a variable rates analysis in BayesTraits
and produces summarized output. Will only work with VR log files
(.VarRates.txt) as directly output from BayesTraits models 1-4, 7 and 9.
Will not (currently) work with output from any other program or model.
}
