% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trees.R
\name{tabulatetree}
\alias{tabulatetree}
\title{Tabulate branch-level information for a tree.}
\usage{
tabulatetree(tree)
}
\arguments{
\item{tree}{a phylogenetic tree as an object of class "phylo".}
}
\value{
A data.frame with n rows equal to the number of branches in the tree
and the following columns:
\itemize{
\item AncNode: the label for the node defining the start of the branch
\item DescNode: the label for the node defining the end of the branch
\item Rbranch: the branch ID
\item length: the branch length
\item des: The list of taxa descending from AncNode
}
}
\description{
Gets all information associated with every branch of a tree.
}
\keyword{internal}
