% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teletubby}
\alias{teletubby}
\title{Create teletubby plots for data distributions.}
\usage{
teletubby(x, col = "indianred", alpha = 0.1, log = T, plot = T)
}
\arguments{
\item{x}{A matrix or data.frame of dimension i>1,j>2 with the values of interest.}

\item{col}{A named R colour or hexadecimal value that defines the colour of the polygon to be plotted (or saved).}

\item{alpha}{The transparency value of the density distribution - for many columns, it might be useful to set this to a low value. Defaults to 0.1}

\item{log}{Boolean operator defining whether or not to log the data provided before calculating the density distribution. Defaults to true.}

\item{plot}{Boolean operator defining whether or not to generate a new plot for the current distributions. Defaults to true. If this is set to false, the function can be used to overlay new polygons onto a pre-existing plot.}
}
\value{
A list of densities (one for each row in the original provided data.frame or matrix).
}
\description{
This function will plot a bunch of DENSITY distributions given a series of ROWS of data.
For each row in the provided matrix / data frame, this function will calculate a density distribution.
Optionally, it will generate polygons to add to a new or pre-existing plot.
}
