% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{tidyVRblock}
\alias{tidyVRblock}
\title{Format VR block from BayesTraits log file}
\usage{
tidyVRblock(VRblock, colstart)
}
\arguments{
\item{VRblock}{Text string including all rows of the VR output relevant to the
specification and definition of rate scalars. Created within the \code{readVR}
function.}

\item{colstart}{Numeric value specifying the column number where scalar information
begins to be recorded in the VRblock output. This varies between different types
of VR model and is internally identified within the \code{readVR} function.}
}
\value{
A \code{data.frame} recording detailed information about the rate
scalars acting at each iteration of the variable rates model.
}
\description{
Helper function used within readVR to format the VR block of a
variable rates output file into an R-friendly data.frame.
}
