% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeUtils.R
\name{tipDescendants}
\alias{tipDescendants}
\title{Identify descendants of a node as named tips}
\usage{
tipDescendants(tree, node)
}
\arguments{
\item{tree}{a phylogenetic tree as an object of class "phylo".}

\item{node}{an integer specifying a node number in the tree.}
}
\value{
The set of named tips descended from node in a vector.
}
\description{
Wrapper for phytools function getDescendants that retrieves tip labels rather than just node labels.
}
