% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trees.R
\name{tippytomies}
\alias{tippytomies}
\title{Identify tippytomies}
\usage{
tippytomies(tre, cutoff = 0.1, action = "identify")
}
\arguments{
\item{cutoff}{The branch length below which taxa will be identified / removed.}

\item{action}{One of two character inputs: "identify" or "delete". See Returns for Details.}

\item{tree}{A phylogenetic tree of class \code{phylo}.}
}
\value{
If "identify", the function will return a list of taxa that belong to each tippytomy. If "delete", the list is returned along with a tree which has deleted all but one (the first) member of each tippytomy. This tree can be used in subsequent comparative analyses.
}
\description{
A function that identifies (and optionally deletes) 'tippytomies' from a tree. A tippytomy refers to the situation in which a pair (or group) of taxa have very short terminal branches resulting in what is essentially a bifurcation or polytomy at the tips of the tree. This can cause numerical issues in several programs. For instance, if two sister taxa have very short branches but very different trait values, the variable rates model will capitalize on this by maximizing the rate of evolution on these branches.
}
