% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{trimBTlog}
\alias{trimBTlog}
\title{Trim header from BayesTraits log files.}
\usage{
trimBTlog(
  dir = ".",
  out = ".",
  pat = ".txt.Log.txt",
  start = "\\tTree No",
  tail = NA,
  resample = NA,
  files = NULL
)
}
\arguments{
\item{dir}{Directory of files for which we want trimmed output. Defaults to the current working directory.}

\item{out}{The name of the folder to which we want output saved to. Defaults to the current working directory.}

\item{pat}{If no files are specified, the function will use the string defined by pat to search for files in dir. If only one or a few specific files are desired, then see the \code{files} parameter.}

\item{start}{A unique search string that defines the start of the columns of output (header row).
Defaults to Tree No which should work for the models mentioned above - check if testing other BayesTraits outputs.}

\item{tail}{Takes a single numeric value.
If defined, tail defines the number of iterations to be retained from the END of the output log.
Defaults to NA, which will return all rows in the original log file (though see the resample parameter)
If tail = 1000, the output will return only the last 1000 rows.}

\item{resample}{Takes a single numeric value.
If defined, the function will resample this number of rows from the original log file.
Note that this number must be greater than the total number of rows in the original file, or this will fail.}

\item{files}{Takes a string or list of strings defining specific log files to be trimmed.
If this is defined, \code{dir} and \code{pat} are ignored.}
}
\value{
A \code{list} containing the trimmed log file, each as a \code{data.frame}.
}
\description{
Removes the header from BayesTraits output files.
Can only work with log files (.txt.Log.txt) as directly output from BayesTraits models 1-4, 7 and 9.
Will not work with output from any other program or model.
}
