% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiags.R
\name{ggAutoCor}
\alias{ggAutoCor}
\title{ggAutoCor}
\usage{
ggAutoCor(posterior, param, max.lag = NULL, min.lag = 0, title)
}
\arguments{
\item{param}{The parameter to do the autocorrelation plot for.}

\item{max.lag}{Maximum lag}

\item{min.lag}{Minimum lag}

\item{psoterior}{The data for the plot, in an object of class vector.}

\item{conf}{The confidence level you want to see the limits for.}
}
\description{
Makes an autocorrealtion plot using ggplot2 - which looks nicer than the
core R plot functions.
}
