% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Marsupial_tree}
\alias{Marsupial_tree}
\title{Example Tree: Marsupial_tree}
\format{
\subsection{\code{Marsupial_tree}}{

A \code{Phylo} format phylogenetic tree of marsupials with 246 species:
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Marsupial_tree
}
\description{
A phylogeny of 246 marsupial species.
}
\keyword{datasets}
