% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Marsupials}
\alias{Marsupials}
\title{Example Dataset: Marsupials
An example continuous data file.}
\format{
\subsection{\code{Marsupials}}{

A \code{data.frame} of marsupial body size.
\describe{
\item{tip_label}{Species name.}
\item{Body}{body size data}
}
}
}
\source{
BayesTraits v4.1 example files.
}
\usage{
Marsupials
}
\description{
Example Dataset: Marsupials
An example continuous data file.
}
\keyword{datasets}
