% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{predBTlog}
\alias{predBTlog}
\title{Calculate predictions from BayesTraits regressions.}
\usage{
predBTlog(input)

predBTlog(input)
}
\arguments{
\item{input}{A character vector defining the input file exactly as output by BayesTraits.}

\item{output}{an optional argument. If defined, a character vector locating the original output file (unmodified).
If unspecified, the function will search for the original input file as if it had been run through BayesTraits, appending .Log.txt.}

\item{plot}{Optional argument. If TRUE, the function will generate a prediction plot. For complicated datasets this might produce undesirable results, and so the behaviour defaults to FALSE.
This should really only be used in the case of a simple linear regression (i.e. a single continuous predictor and a single continuous response).}
}
\value{
A matrix of predicted values for terminal taxa (if original input was MCMC, one iteration per column)

A numeric value defining the log marginal likelihood of the output model.
}
\description{
Function to calculate predictions from BayesTraits input/output files. Incompatible with other programs.

Function to extract the marginal likelihood from the raw stepping-stone sampling output.
}
