% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTLogs.R
\name{readBTlog}
\alias{readBTlog}
\title{Read in BayesTraits Log Files (raw)}
\usage{
readBTlog(file, burnin = 0)
}
\arguments{
\item{file}{Takes a single string defining the direct path to a single log file.}

\item{burnin}{Value specifying the number of samples to remove from the beginning of the log file before summarizing.
The default value is zero (i.e. use the full output of the sampled chain).}
}
\value{
R data.frame object that contain the trimmed log file.
}
\description{
Removes the header from BayesTraits output files.
Can only work with log files (.txt.Log.txt) as directly output from BayesTraits models 1-4, 7 and 9.
Will not work with output from any other program or model.
}
