% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppBTVR.R
\name{summarizeVR}
\alias{summarizeVR}
\title{Summarize BayesTraits VR Log Files (from raw output)}
\usage{
summarizeVR(vrfile, tree, forcedequaltrees = F, burnin = 0)
}
\arguments{
\item{vrfile}{Takes a single string defining the direct path to a single variable rates log file.}

\item{tree}{A phylogenetic tree object of class \code{phylo}. The output will link rates to the given tree. That is, every rate scalar is defined on the basis of a branch. If that branch does not exist in the defined tree, the rate scalar will instead be applied to the branch which terminates in the most recent common ancestor of all taxa that descended from the original defined branch. This is designed to summarize the results of a multi-topology variable rates output. The input tree must contain all taxa in the original tree used to run the variable rates model.}

\item{forcedequaltrees}{Boolean operator, default = \code{FALSE}. If \code{forcedequaltrees = true}, then the model will be run iteratively over a pre-defined sample of trees.In this situation, we will have duplicate iterations (one for each tree) and thus multiple posteriors. We need to let the post-processor know if this is the case.}

\item{burnin}{A value specifying the number of rows to remove from the beginning of each log file. Default value is 0 and so the entire chain will be used to calculate rates.}
}
\value{
A list with three elements:
\itemize{
\item summary - A \code{data.frame} giving information about each branch/node
in the analysis. Each branch is identified by the numeric value in "branch". For
samples of trees, this value is equivalent to the "partition ID" used internally
by BayesTraits. "NoDes" and "Des" record the number and names of all descendant
taxa from each branch. The "freq" and "probability" columns tell us the frequency
and probability that this branch occur in the sample are only relevant for
analyses over samples of trees. The "length" records the branch length and
is left blank for tree samples. The "DesBranches" tells us all branches that descend from
the branch of interest. The median and mean scalar are the median and mean rate
scalars acting on that branch across the entire posterior distribution. n_scaled
tells us how many iterations that branch had a scalar acting on it, and n_origin tells
us how many iterations that branch had a scalar ORIGINATING on it (either as the base of
a node scalar or as its own branch scalar).
\item posterior - A \code{data.frame} providing the posterior distribution of rate scalars acting
on every branch (column) at each iteration (row).
\item tree - The input tree will be returned for use in tree-scaling and visualization etc.
}
}
\description{
Reads the output file from a variable rates analysis in BayesTraits
and produces summarized output. Will only work with VR log files
(.VarRates.txt) as directly output from BayesTraits models 1-4, 7 and 9.
Will not (currently) work with output from any other program or model.
}
\examples{
\dontrun{
# Here we define a file based on one of the examples included with this package
VRlogfilepath <- system.file("extdata", "MammalBody_VR-001.txt.VarRates.txt")
summarizeVR(vrfile = VRlogfilepath, tree = Mammal_trees)
}
}
